M6 step 2: module wiring via platform.Deps + Register hooks
Introduce the module contract from docs/M6-plan.md §6.1 so tools (step 3) and
later heatmap/sync can each mount themselves uniformly:

- platform.Deps bundles the shared infra a module registers against (Router,
  LoginPage, DB, Cal, Google); its doc states the contract (RegisterRoutes at
  setup, RegisterPages at serve), the /api/ auth-by-prefix rule, and single
  ownership of the /* catch-all.
- app now exposes package-level RegisterRoutes(deps) (API routes + the
  outside-/api nudge) and RegisterPages(deps) (the /tools and /* pages); the
  old (*Server).RegisterRoutes method becomes unexported registerAPI.
- The homePage/toolsPage HTML consts and their render handlers move verbatim
  from main.go into internal/app/pages.go.
- main.go builds Deps once and calls the module hooks; the appSrv global, store
  construction, nudge registration, and inline page handlers are gone.

Behavior-preserving: endpoints and pages unchanged (verified with a real serve —
/, /tools, /api/status all 200; full ./dev ci green). No data-model changes.
